mcp_server/pyproject.toml
[project]
name = "crucible-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0",
    "httpx>=0.27",
    "pydantic-settings>=2.6",
    "pyyaml>=6.0",
    "uvicorn[standard]>=0.32.0",
    "crucible-lib",
]

[tool.uv.sources]
crucible-lib = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crucible_mcp"]
